change returned type of variable botState from Long to int

+++ src/main/java/com/resliv/turbot/model/User.java
@@ -20,7 +20,7 @@ public class User {
     private Long id;
     private Long chatId;
     private String name;
-    private Long botState;
+    private int botState;
 
 
     @ManyToOne
